🛠️ Simplify auth setup and config structure
