+++ src/main/java/com/store/book/config/OpenApiConfig.java
@@ -1,0 +1,30 @@
+package com.store.book.config;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenApiConfig {
+    @Bean
+    public OpenAPI customOpenApi() {
+        final String securitySchemeName = "basicAuth";
+        return new OpenAPI()
+                .info(new Info()
+                              .title("Bookstore API")
+                              .version("1.0")
+                              .description("API for managing users and books"))
+                .addSecurityItem(new SecurityRequirement().addList(securitySchemeName))
+                .components(new Components().addSecuritySchemes(
+                        securitySchemeName,
+                        new SecurityScheme()
+                                .name(securitySchemeName)
+                                .type(SecurityScheme.Type.HTTP)
+                                .scheme("basic")
+                                                               ));
+    }
+}
