+++ src/main/java/com/store/book/config/SecurityConfig.java
@@ -1,10 +1,5 @@
 package com.store.book.config;
 
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -25,6 +20,11 @@ public class SecurityConfig {
     private final UserDetailsService userDetailsService;
 
     @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http
                 .csrf(AbstractHttpConfigurer::disable)
@@ -39,24 +39,5 @@ public class SecurityConfig {
                         .anyRequest().authenticated())
                 .httpBasic(Customizer.withDefaults());
         return http.build();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public OpenAPI customOpenApi() {
-        final String securitySchemeName = "basicAuth";
-        return new OpenAPI().info(new Info().title("Bookstore API").version("1.0")
-                                            .description("API for managing users and books"))
-                            .addSecurityItem(new SecurityRequirement().addList(securitySchemeName))
-                            .components(new Components().addSecuritySchemes(
-                                    securitySchemeName,
-                                    new SecurityScheme().name(securitySchemeName)
-                                                        .type(SecurityScheme.Type.HTTP)
-                                                        .scheme("basic")
-                                                                           ));
     }
 }
