+++ src/main/java/com/store/book/model/RoleName.java
@@ -3,4 +3,3 @@ package com.store.book.model;
 public enum RoleName {
     USER, ADMIN
 }
-
